pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-server"
version = "0.1.0"
description = "a server for rendering markdown files"
readme = "README.md"
authors = [
    {name = "Vinzent Brömauer", email = "vinz.corno@web.com"}
]
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "markdown-it-py",
    "flask",
    "markupsafe",
    "latex2mathml"
]

[project.scripts]
markdown-server = "markdown_server.server:main"
